remove old tpp-5.2-fix.diff, install-R-packages.R
* notes.org (Tangling):
(Xtandem): tpp-5.2-fix.diff is no longer needed, because we use the comet shipped with tpp source
and this fix was for an older tpp, so this is no longer relevant.
This patch file was  used in the dockerifle of the python implementation of gladiator,
but since we no longer need that, we can remove it
(Dependencies): no longer tangle install-R-packages.R,
because we don't use the above dockerfile anymore, where this was used on to install R packages.
Instead these guix has been responsible for this.